Generate cidrs.txt from a YAML source version
Because the YAML source can include comments about what the CIDRs actualy are, which is useful to understand what is and isn't included.
Couldn't put comments in cidrs.txt because that is in use by repos already, and comments would disrupt those clients.
YAML is also helpful when we want different allowlists for different apps, as seen in https://github.com/ministryofjustice/hmpps-ip-allowlists/tree/main